separate rsd into its own logic, add new docs, refactor trackback implementation
