+++ docs/source/rsd.rst
@@ -1,0 +1,6 @@
+Really Simple Discovery (RSD)
+============================
+
+IndieWeb Utils contains an `rsd_discovery` function for use in finding the values associated with keys in an `EditURI` `Really Simple Discovery <https://en.wikipedia.org/wiki/Really_Simple_Discovery>`_ document.
+
+.. autofunction:: indieweb_utils.rsd_discovery
