+++ docs/source/trackback.rst
@@ -1,7 +1,7 @@
 Trackbacks
 ==========
 
-This library includes support for Trackback URL Discovery and sending Trackbacks.
+This library includes support for `Trackback <http://archive.cweiske.de/trackback/trackback-1.2.html>`_ URL Discovery and sending Trackbacks.
 
 Trackback URL discovery
 -----------------------
@@ -11,12 +11,7 @@ The IndieWeb Utils Trackback discovery performs the following steps:
 1. Checks for a `trackback:ping` attribute in an RDF comment. If one is found, this is returned.
 2. Checks for an `EditURI` `<link>` tag. If one is found, its contents are retrieved. If the contents contain a `trackback:ping` attribute, this is returned.
 
-.. code-block:: python
-
-    from indieweb_utils import discover_trackback_url
-
-    discover_trackback_url('http://example.com/post/123')
-
+.. autofunction:: indieweb_utils.trackback.discover_trackback_url
 
 Send a Trackback
 ----------------
