+++ src/indieweb_utils/rsd/__init__.py
@@ -1,0 +1,3 @@
+from discover import rsd_discovery
+
+__all__ = ["rsd_discovery"]
